refactor: step two 페이지 컴포넌트 책임 분리
